Move the check module function to split file
