+++ cmd/db-exporter/src/main.rs
@@ -1,13 +1,14 @@
 // Copyright (c) The Starcoin Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
+
 use anyhow::{bail, format_err, Result};
 use bcs_ext::BCSCodec;
 use bcs_ext::Sample;
 use clap::IntoApp;
 use clap::Parser;
 use csv::Writer;
-use db_exporter::verify_modules_via_export_file;
+use db_exporter::verify_module::verify_modules_via_export_file;
 use indicatif::{MultiProgress, ProgressBar, ProgressStyle};
 use serde::ser::SerializeMap;
 use serde::{Serialize, Serializer};
@@ -69,6 +70,7 @@ use std::sync::atomic::{AtomicU64, Ordering};
 use std::sync::Arc;
 use std::time::SystemTime;
 use std::{thread, thread::JoinHandle};
+
 
 const BLOCK_GAP: u64 = 1000;
 const BACK_SIZE: u64 = 10000;
@@ -314,6 +316,7 @@ pub struct StartupInfoBackOptions {
     #[clap(long, short = 'b')]
     pub back_size: Option<u64>,
 }
+
 #[derive(Debug, Copy, Clone)]
 pub enum Txntype {
     CreateAccount,
